fix: breakage report submission
