+++ shared/js/background/broken-site-report.js
@@ -1,3 +1,5 @@
+import { getTrackerAggregationStats } from '../ui/models/mixins/calculate-aggregation-stats'
+
 /**
  *
  * This is part of our tool for anonymous broken site reports
@@ -67,4 +69,57 @@ export function getURL (pixelName) {
 
     const url = 'https://improving.duckduckgo.com/t/'
     return url + pixelName
+}
+
+/**
+ * Given an optional category and description, create a report for a given Tab instance.
+ *
+ * This code previously lived within the UI section of the dashboard,
+ * but has been moved here since there's no longer a relationship to 'where' this request
+ * came from.
+ *
+ * @param {import("./classes/tab.es6")} tab
+ * @param {string | undefined} category - optional category
+ * @param {string | undefined} description - optional description
+ */
+export function breakageReportForTab (tab, category, description) {
+    /**
+     * Returns a list of tracker URLs after looping through all the entities.
+     * @param {import('../ui/models/mixins/calculate-aggregation-stats.js').AggregateCompanyData[]} list
+     * @returns {string[]}
+     */
+    function collectAllUrls (list) {
+        const urls = []
+        list.forEach(item => {
+            item.urlsMap.forEach((_, url) => urls.push(url))
+        })
+        return urls
+    }
+
+    const siteUrl = tab.url?.split('?')[0].split('#')[0]
+    if (!siteUrl) {
+        return
+    }
+    const aggregationStats = getTrackerAggregationStats(tab.trackers)
+    const blockedTrackers = collectAllUrls(aggregationStats.blockAction.list)
+    const surrogates = collectAllUrls(aggregationStats.redirectAction.list)
+    const urlParametersRemoved = tab.urlParametersRemoved ? 'true' : 'false'
+    const ampUrl = tab.ampUrl || undefined
+    const upgradedHttps = tab.upgradedHttps
+    const tdsETag = settings.getSetting('tds-etag')
+
+    const brokenSiteParams = new URLSearchParams({
+        siteUrl,
+        upgradedHttps: upgradedHttps.toString(),
+        tds: tdsETag,
+        urlParametersRemoved,
+        blockedTrackers: blockedTrackers.join(','),
+        surrogates: surrogates.join(',')
+    })
+
+    if (ampUrl) brokenSiteParams.set('ampUrl', ampUrl)
+    if (category) brokenSiteParams.set('category', category)
+    if (description) brokenSiteParams.set('description', description)
+
+    return fire(brokenSiteParams.toString())
 }
